+++ backend/interfaces/api/server/handler/top_page.go
@@ -8,7 +8,7 @@ import (
 	"github.com/yoshihiro-shu/draft-backend/backend/domain/model"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/cache"
-	"github.com/yoshihiro-shu/draft-backend/backend/internal/pkg/logger"
+	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
 	"github.com/yoshihiro-shu/draft-backend/backend/internal/pkg/pager"
 	"go.uber.org/zap"
 )
