+++ backend/registory/latest_articles_repository.go
@@ -5,7 +5,7 @@ import (
 	"github.com/yoshihiro-shu/draft-backend/backend/infrastructure/persistence"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/handler"
 	"github.com/yoshihiro-shu/draft-backend/backend/interfaces/api/server/request"
-	"github.com/yoshihiro-shu/draft-backend/backend/internal/pkg/logger"
+	"github.com/yoshihiro-shu/draft-backend/backend/internal/logger"
 	"gorm.io/gorm"
 )
